Move utils in to the validators.
